+++ R/pal-hue.r
@@ -36,6 +36,8 @@ hue_pal <- function(h = c(0, 360) + 15, c = 100, l = 65, h.start = 0, direction 
     rotate <- function(x) (x + h.start) %% 360 * direction
     hues <- rotate(seq(h[1], h[2], length.out = n))
 
-    grDevices::hcl(hues, c, l)
+    lch <- cbind(l, c, hues)
+    rgb <- farver::convert_colour(lch, "lch", "rgb")
+    rgb2col(rgb)
   }
 }