+++ R/scales-package.R
@@ -1,6 +1,4 @@
 #' @keywords internal
-#' @useDynLib scales
-#' @importFrom Rcpp evalCpp
 "_PACKAGE"
 
 # The following block is used by usethis to automatically manage