+++ tests/testthat/test-colour-manip.r
@@ -1,22 +1,26 @@
-context("Alpha")
+# hcl ---------------------------------------------------------------------
 
-hex <- function(x) {
-  rgb <- col2rgb(x, TRUE) / 255
-  rgb(rgb[1, ], rgb[2, ], rgb[3, ], rgb[4, ])
-}
+test_that("can modify each hcl component", {
+  expect_equal(col2hcl("red", h = 180), "#00A17D")
+  expect_equal(col2hcl("red", l = 50), "#F40000")
+  expect_equal(col2hcl("red", c = 50), "#C7624A")
+  expect_equal(col2hcl("red", alpha = 0.5), "#FF000080")
+})
+
+# alpha -------------------------------------------------------------------
 
 test_that("missing alpha preserves existing", {
-  cols <- col2rgb(rep("red", 5), TRUE) / 255
-  cols[4, ] <- seq(0, 1, length.out = ncol(cols))
+  cols <- col2rgb(rep("red", 5), TRUE)
+  cols[, 4] <- seq(0, 255, length.out = nrow(cols))
 
-  reds <- rgb(cols[1, ], cols[2, ], cols[3, ], cols[4, ])
+  reds <- rgb2col(cols)
 
   expect_equal(reds, alpha(reds, NA))
   expect_equal(reds, alpha(reds, rep(NA, 5)))
 })
 
 test_that("alpha values recycled to match colour", {
-  cols <- hex(c("red", "green", "blue", "pink"))
+  cols <- rgb2col(col2rgb(c("red", "green", "blue", "pink")))
 
   expect_equal(cols, alpha(cols, NA))
   expect_equal(cols, alpha(cols, 1))
@@ -25,7 +29,7 @@ test_that("alpha values recycled to match colour", {
 test_that("col values recycled to match alpha", {
   alphas <- round(seq(0, 255, length.out = 3))
   reds <- alpha("red", alphas / 255)
-  reds_alpha <- col2rgb(reds, TRUE)[4, ]
+  reds_alpha <- col2rgb(reds, TRUE)[, 4]
 
   expect_equal(alphas, reds_alpha)
 })