+++ tests/testthat/test-colour-ramp.R
@@ -10,21 +10,21 @@ test_that("Special values yield NAs", {
 })
 
 test_that("Fully opaque colors are returned without alpha", {
-  expect_identical(
+  expect_equal(
     colour_ramp(c("#1234AB", "#BA4321"))(0:1),
     c("#1234AB", "#BA4321")
   )
 })
 
 test_that("Partially transparent colors are returned with alpha", {
-  expect_identical(
+  expect_equal(
     colour_ramp(c("#1234AB20", "#BA43218F"))(0:1),
     c("#1234AB20", "#BA43218F")
   )
 })
 
 test_that("Partially transparent colors are returned without alpha when `alpha = FALSE`", {
-  expect_identical(
+  expect_equal(
     colour_ramp(c("#1234AB20", "#BA43218F"), alpha = FALSE)(0:1),
     c("#1234AB", "#BA4321")
   )